requirements.txt
albumentations==1.4.15
numpy==1.24.4
numpy==1.19.5
opencv_python==4.3.0.36
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==8.1.0
Pillow==11.0.0
pytz==2020.5
pytz==2023.4
PyYAML==6.0.2
PyYAML==5.4.1
PyYAML==6.0.1
PyYAML==6.0.2
timm==0.5.4
torch==1.11.0+cu113
torch==2.4.1+cu124
torchvision==0.12.0+cu113
torchvision==0.19.1+cu124
typed_argument_parser==1.10.1
ultralytics==8.3.21
